Api-Chat/requirements.txt
flask
python-dotenv
openai
flask-cors
requests
black
Flask[async]
werkzeug